refactor: move script logic to app modules
