+++ src/app/io/__init__.py
@@ -1,0 +1,3 @@
+"""I/O helpers for the application."""
+
+__all__ = ["csvio"]
