+++ src/app/utils/__init__.py
@@ -1,0 +1,3 @@
+"""Utility functions for data processing."""
+
+__all__ = ["mac", "notes"]
